Fix simulate pipeline API
The simulate pipeline API uses a special deep copy method to write out
its content via toXContent - this method does not support writing out
sets, resulting in an error as this processor writes out data using
sets.

This commit fixes this issue by replacing the set with a list, so that
the serialization works again, at the small price of creating more
objects when running.

Closes #27